requirements.txt
# Core FastAPI and async dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP client and LLM integration
httpx==0.25.2
openai==1.3.7

# Vector embeddings and ML
sentence-transformers==2.2.2
numpy>=1.26.0

# Speech processing
openai-whisper==20231117

# Monitoring and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
prometheus-client==0.19.0

# Rate limiting and circuit breakers
pybreaker==1.0.1
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Development and linting
ruff==0.1.6
mypy==1.7.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.2.0
python-dateutil==2.8.2
